// isomatic-frontend/src/app/shared/_generated/rest-api/model/create-device-model-request.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DeviceFunctionBlueprint } from './device-function-blueprint';


export interface CreateDeviceModelRequest { 
    manufacturerAbbr?: string;
    manufacturerName?: string;
    name?: string;
    type?: string;
    functionBlueprints?: Array<DeviceFunctionBlueprint>;
}