// isomatic-frontend/src/app/shared/_generated/rest-api/model/device-unit-dto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DeviceUnitSerialNumber } from './device-unit-serial-number';
import { DeviceModelDto } from './device-model-dto';
import { DeviceUnitRegistrationStatus } from './device-unit-registration-status';


export interface DeviceUnitDto { 
    registrationStatus?: DeviceUnitRegistrationStatus;
    serialNumber?: DeviceUnitSerialNumber;
    modelDto?: DeviceModelDto;
}